Refactor folder structure and cors config
