+++ packages/api/src/app.ts
@@ -1,38 +1,54 @@
+import { clerkMiddleware } from "@hono/clerk-auth";
+import { trpcServer } from "@hono/trpc-server";
 import { Hono } from "hono";
 import { cors } from "hono/cors";
 
+import { connect } from "@knighthacks/db";
+
 import type { HonoConfig } from "./config";
-import { clerk } from "./middlewares/clerk";
-import { db } from "./middlewares/db";
-import { trpc } from "./middlewares/trpc";
 import { resume } from "./routes/resume";
 import { webhook } from "./routes/webhook";
+import { createTRPCContextFromHonoContext } from "./trpc/context";
+import { appRouter } from "./trpc/routers";
 
 const app = new Hono<HonoConfig>()
   .get("/", (c) => {
     return c.text("Hello world");
   })
-  .use(
-    "*",
-    cors({
-      origin: [
-        // Admin tool dev server
-        "http://localhost:3000",
-        // Club website dev server
-        "http://localhost:3001",
-        // Knight Hacks 2024 dev server
-        "http://localhost:3002",
-        // Admin tool production server
-        "https://knighthacks-admin.pages.dev",
-        "https://admin.knighthacks.org",
-        "https://2024.knighthacks.org",
-        "https://knighthacks.org",
-      ],
-    }),
-  )
-  .use("*", db)
-  .use("*", clerk)
-  .use("/trpc/*", trpc)
+  .use("*", (c, next) => {
+    const origin =
+      c.env.ENV === "development"
+        ? "*"
+        : [
+            "https://admin.knighthacks.org",
+            "https://2024.knighthacks.org",
+            "https://club.knighthacks.org",
+            "https://knighthacks.org",
+          ];
+    return cors({
+      origin,
+    })(c, next);
+  })
+  .use("*", (c, next) => {
+    const db = connect(c.env.DATABASE_URL, c.env.DATABASE_AUTH_TOKEN);
+    c.set("db", db);
+    return next();
+  })
+  .use("*", (c, next) => {
+    return clerkMiddleware({
+      publishableKey: c.env.CLERK_PUBLISHABLE_KEY,
+      secretKey: c.env.CLERK_SECRET_KEY,
+    })(c, next);
+  })
+  .use("/trpc/*", (c, next) => {
+    return trpcServer({
+      router: appRouter,
+      onError({ error }) {
+        console.error(error);
+      },
+      createContext: createTRPCContextFromHonoContext(c),
+    })(c, next);
+  })
   .route("/resume", resume)
   .route("/webhook", webhook);
