--- packages/api/src/middlewares/index.ts
@@ -1,3 +1,0 @@
-export * from "./clerk";
-export * from "./db";
-export * from "./trpc";
